+++ modules/client/CHANGELOG.md
@@ -15,6 +15,8 @@ TEMPLATE:
 ## [UPCOMING]
 ### Added
 - Add ipfs optimization.
+### Fixed
+- Fix plugin installation
 ## [1.3.1]
 ### Fixed
 - Polygon network names